Create CombineAllTXTFiles
This batch file will loop through all files ending in .txt in the directory this batch file is located in. It will store the contents of each file in an output file called cominedresults.txt. If the output file already exists, it will be overwritten. The file will be stored in a subdirectory of the directory this batch file is located in called combinedresults. If the subdirectory doesn't exist it will be created. This batch file will execute the quickest if run from the local computer on which it resides, but running it from a UNC path is possible.