Add variables and reorganize
Allowed user to change the value of the file extension that is searched for. Also separate variables into those that the user should update before running and those that shouldn't be edited. Lastly, saved the full path of the output file to a variable so it could be referenced everywhere it is needed.